# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fall Detection API",
        "description": "API documentation for the Fall Detection System",
        "contact": {
            "email": "admin@fall-detection.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://fall-detection-backend.test/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/elderly/register": {
            "post": {
                "tags": [
                    "Elderly Authentication"
                ],
                "summary": "Register a new elderly user",
                "description": "Create a new elderly user account with profile information",
                "operationId": "elderlyRegister",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "device_name",
                                    "first_name",
                                    "last_name",
                                    "date_of_birth",
                                    "gender",
                                    "primary_phone",
                                    "current_address",
                                    "emergency_contact_name",
                                    "emergency_contact_phone",
                                    "emergency_contact_relationship",
                                    "mobility_status",
                                    "vision_status",
                                    "hearing_status",
                                    "care_level",
                                    "preferred_language",
                                    "device_status",
                                    "living_situation",
                                    "activity_level"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "device_name": {
                                        "type": "string",
                                        "example": "John's iPhone"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1940-01-01"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female",
                                            "other"
                                        ],
                                        "example": "male"
                                    },
                                    "primary_phone": {
                                        "type": "string",
                                        "example": "+1234567890"
                                    },
                                    "current_address": {
                                        "type": "string",
                                        "example": "123 Main St"
                                    },
                                    "emergency_contact_name": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "emergency_contact_phone": {
                                        "type": "string",
                                        "example": "+1987654321"
                                    },
                                    "emergency_contact_relationship": {
                                        "type": "string",
                                        "example": "Daughter"
                                    },
                                    "mobility_status": {
                                        "type": "string",
                                        "enum": [
                                            "independent",
                                            "needs_assistance",
                                            "wheelchair_bound"
                                        ],
                                        "example": "independent"
                                    },
                                    "vision_status": {
                                        "type": "string",
                                        "enum": [
                                            "normal",
                                            "glasses",
                                            "impaired"
                                        ],
                                        "example": "normal"
                                    },
                                    "hearing_status": {
                                        "type": "string",
                                        "enum": [
                                            "normal",
                                            "hearing_aid",
                                            "impaired"
                                        ],
                                        "example": "normal"
                                    },
                                    "care_level": {
                                        "type": "string",
                                        "enum": [
                                            "basic",
                                            "moderate",
                                            "intensive"
                                        ],
                                        "example": "basic"
                                    },
                                    "preferred_language": {
                                        "type": "string",
                                        "example": "English"
                                    },
                                    "device_status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive"
                                        ],
                                        "example": "active"
                                    },
                                    "living_situation": {
                                        "type": "string",
                                        "enum": [
                                            "lives_alone",
                                            "with_family",
                                            "assisted_living"
                                        ],
                                        "example": "with_family"
                                    },
                                    "activity_level": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "moderate",
                                            "sedentary"
                                        ],
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "1|abcdefghijklmnopqrstuvwxyz"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The email has already been taken."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/elderly/login": {
            "post": {
                "tags": [
                    "Elderly Authentication"
                ],
                "summary": "Login an elderly user",
                "description": "Authenticate an elderly user and return an access token",
                "operationId": "elderlyLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "device_name"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "device_name": {
                                        "type": "string",
                                        "example": "John's iPhone"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "1|abcdefghijklmnopqrstuvwxyz"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The provided credentials are incorrect."
                                        },
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The provided credentials are incorrect."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/elderly/logout": {
            "post": {
                "tags": [
                    "Elderly Authentication"
                ],
                "summary": "Logout an elderly user",
                "description": "Invalidate the current access token",
                "operationId": "elderlyLogout",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/elderly/user": {
            "get": {
                "tags": [
                    "Elderly Authentication"
                ],
                "summary": "Get authenticated elderly user",
                "description": "Get the currently authenticated elderly user's information",
                "operationId": "elderlyUser",
                "responses": {
                    "200": {
                        "description": "User information retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "required": [
                    "name",
                    "email",
                    "role"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "admin",
                            "carer",
                            "elderly"
                        ],
                        "example": "elderly"
                    },
                    "device_token": {
                        "type": "string",
                        "example": "fcm_token_here",
                        "nullable": true
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "elderly_profile": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/ElderlyProfile"
                            }
                        ],
                        "nullable": true
                    },
                    "carer_profile": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/CarerProfile"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}